+++ commandline/Albatross.CommandLine/OptionAttribute.cs
@@ -10,6 +10,9 @@ namespace Albatross.CommandLine {
 		public bool Hidden { get; set; }
 		public bool Required { get; set; }
 		public bool Ignore { get; set; }
+		/// <summary>
+		/// When true, the code generator will generate a default value based on the initializer of the property
+		/// </summary>
 		public bool DefaultToInitializer { get; set; }
 	}
 	public class ArgumentAttribute : Attribute {
@@ -20,6 +23,9 @@ namespace Albatross.CommandLine {
 		public string Name { get; set; }
 		public string? Description { get; set; }
 		public bool Hidden { get; set; }
+		/// <summary>
+		/// When true, the code generator will generate a default value based on the initializer of the property
+		/// </summary>
 		public bool DefaultToInitializer { get; set; }
 	}
 }